requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
python-dotenv==1.0.1
twilio==9.0.4
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
python-multipart==0.0.9
